# Remove allocations through interning and speedup hashing