```
==>  ./flamegraph/test/perf-vertx-stacks-01.txt  <==
Benchmark #1: target/release/inferno-collapse-xdebug ./flamegraph/test/perf-vertx-stacks-01.txt
  Time (mean ± σ):       1.8 ms ±   0.2 ms    [User: 1.5 ms, System: 0.8 ms]
  Range (min … max):     1.5 ms …   3.3 ms    759 runs
 
  Warning: Command took less than 5 ms to complete. Results might be inaccurate.
 
Benchmark #2: ./flamegraph/stackcollapse-perf.pl --all ./flamegraph/test/perf-vertx-stacks-01.txt
  Time (mean ± σ):      62.4 ms ±   1.1 ms    [User: 59.6 ms, System: 2.6 ms]
  Range (min … max):    60.2 ms …  64.0 ms    45 runs
 
Summary
  'target/release/inferno-collapse-xdebug ./flamegraph/test/perf-vertx-stacks-01.txt' ran
   34.05 ± 4.29 times faster than './flamegraph/stackcollapse-perf.pl --all ./flamegraph/test/perf-vertx-stacks-01.txt'
```