@HeroicKatora 2019-02-11
Anyone reading this, feel free to comment. Any additional suggestions and improvements, or bugfixes are going to be taken into account tomorrow. @neosilky  I didn't know how to properly test these changes, any suggestsions?