@HeroicKatora 2019-02-11
Closing this until I find the speed tests etc.