@neosilky 2019-02-11
Hi @HeroicKatora!

In terms of testing, I've been running this program against the original `stackcollapse-xdebug.php` and comparing the output. In terms of benchmarking, I've been using `hyperfine` in the following way:

```bash
hyperfine "./target/release/inferno-collapse-xdebug xdebug.trace.xt > /dev/null" "./flamegraph/stackcollapse-xdebug.php xdebug.trace.xt > /dev/null"
```

I'll grab a copy of these changes and see what effect they have.