@neosilky 2019-02-11
Nice one @HeroicKatora! Shall I reopen and we can look at getting this merged?

```
Benchmark #1: ./target/release/inferno-collapse-xdebug xdebug.trace.xt > /dev/null
  Time (mean ± σ):     214.3 ms ±  13.2 ms    [User: 198.3 ms, System: 14.0 ms]
  Range (min … max):   198.9 ms … 237.7 ms    12 runs

Benchmark #2: ./flamegraph/stackcollapse-xdebug.php xdebug.trace.xt > /dev/null
  Time (mean ± σ):      1.010 s ±  0.030 s    [User: 943.4 ms, System: 57.0 ms]
  Range (min … max):    0.967 s …  1.057 s    10 runs

Summary
  './target/release/inferno-collapse-xdebug xdebug.trace.xt > /dev/null' ran
    4.71 ± 0.32 times faster than './flamegraph/stackcollapse-xdebug.php xdebug.trace.xt > /dev/null'
```