@HeroicKatora 2019-02-11
@neo Ah yes, let's reopen this with some more potential improvements.