@HeroicKatora 2019-02-11
The question is mostly: Is this faster than other Rust versions, not so much only a comparison against php. How large is the stack trace btw, I haven't managed to get my hands on one with a reasonable load.