@neosilky 2019-02-11
@HeroicKatora It's definitely the fastest I've seen so far. Current `master` is around 3.7x faster than PHP and this is around 4.7x.

I have a 22MB (~320,000 lines) one that I've been using for the tests. It's from a client's WordPress site but feel free to email me if you want a copy of it.