@HeroicKatora 2019-02-11
@neosilky (this time the correct ping). Probably got another nice speedup?

```
Benchmark #1: target/release/inferno-collapse-xdebug ./flamegraph/test/xdebug-perf-trace.txt
  Time (mean ± σ):     122.2 ms ±   1.6 ms    [User: 115.8 ms, System: 4.9 ms]
  Range (min … max):   120.1 ms … 125.3 ms    24 runs
```