@neosilky 2019-02-12
@HeroicKatora Excellent! Showing as 5x faster than PHP & the fastest I've seen so far.

It all looks good to me so let me know when you feel like you're done and I can merge. Please run `cargo fmt` over it too.

Thanks!