@HeroicKatora 2019-02-12
@neosilky All ready, still feels like it could somehow be more efficient but can't find anything else for now.